% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVPSUM.R
\name{hvpsum}
\alias{hvpsum}
\title{HVPSUM: Genetic Correlation}
\usage{
hvpsum(
  ldsc_env_path = NULL,
  ldsc_exe_path = NULL,
  ld_path = NULL,
  sumstats1 = NULL,
  sumstats2 = NULL,
  h21_data = NULL,
  h22_data = NULL,
  gencov_data = NULL,
  tau,
  intern = TRUE
)
}
\arguments{
\item{ldsc_env_path}{Character string. The path to the environment where LDSC is installed.}

\item{ldsc_exe_path}{Character string. The path to the LDSC executable.}

\item{ld_path}{Character string. The path to the reference LD blocks.}

\item{sumstats1}{Character string. Path to the first summary statistics file (e.g., .sumstats.gz).}

\item{sumstats2}{Character string. Path to the second summary statistics file (e.g., .sumstats.gz).}

\item{h21_data}{Data frame or vector. Precomputed heritability estimates for the first trait.
If \code{NULL}, it will be computed using LDSC if the LDSC paths are provided.}

\item{h22_data}{Data frame or vector. Precomputed heritability estimates for the second trait.
If \code{NULL}, it will be computed using LDSC if the LDSC paths are provided.}

\item{gencov_data}{Data frame or vector. Precomputed genetic covariance estimates.
If \code{NULL}, it will be computed using LDSC if the LDSC paths are provided.}

\item{tau}{Numeric value. The causal effect to adjust the genetic covariance and heritability.}

\item{intern}{Logical. If \code{TRUE}, the LDSC command is run internally.}
}
\value{
A list containing the following:
\item{means}{A named vector of means for heritability, genetic covariance, and correlation estimates.}
\item{se}{A named vector of standard errors for the estimates.}
\item{pvalues}{A named vector of p-values for the estimates.}
}
\description{
This function applies the HVP model to disentangle horizontal pleiotropy effects from vertical pleiotropy using GWAS summary data. It corrects genetic correlation estimates derived from cross-trait LD-score regression (LDSC).
}
\details{
This r package performs the following steps:
\enumerate{
\item If \code{ldsc_env_path}, \code{ldsc_exe_path}, and \code{ld_path} are provided, it runs the LDSC analysis to generate
heritability and genetic covariance estimates based on summary statistics provided.
\item If precomputed data is provided for \code{h21_data}, \code{h22_data}, and \code{gencov_data}, these are used directly.
\item Calculates and returns the estimates, standard errors, and p-values for the corrected and uncorrected parameters.
}
}
