% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HVPSUM.R
\name{hvpsum}
\alias{hvpsum}
\title{HVPSUM: Genetic Correlation}
\usage{
hvpsum(
  ldsc_env_path = NULL,
  ldsc_exe_path = NULL,
  ld_path = NULL,
  sumstats1 = NULL,
  sumstats2 = NULL,
  h21block_path = NULL,
  h22block_path = NULL,
  gen_cov_path = NULL,
  tau,
  intern = TRUE
)
}
\arguments{
\item{ldsc_env_path}{Path to the LDSC environment.}

\item{ldsc_exe_path}{Path to the LDSC executable script.}

\item{ld_path}{Path to the reference LD files.}

\item{sumstats1}{Path to the first trait munged summary statistics file (outcome).}

\item{sumstats2}{Path to the second trait munged summary statistics file(exposure).}

\item{h21block_path}{Path to the pre-computed jackknife blocks estimate  for the heritability of trait 1.}

\item{h22block_path}{Path to the pre-computed jackknife blocks estimate for the heritability of the trait 2.}

\item{gen_cov_path}{Path to the pre-computed jackknife blocks estimate for the genetic covariance between the traits.}

\item{tau}{Causal effect of the exposure on the outcome, as estimated by a robust Mendelian Randomization (MR) method.}

\item{intern}{Logical; if \code{TRUE}, captures the LDSC output internally when running the command.
Defaults to \code{TRUE}.}
}
\value{
A list containing:
\describe{
\item{Estimates}{Estimates for heritabilities, genetic covariance, and genetic correlation (corrected and uncorrected).}
\item{SE}{Standard errors for each estimate.}
\item{p-values}{P-values associated with each estimate.}
}
}
\description{
This function applies the HVP model to disentangle horizontal pleiotropy effects from vertical pleiotropy using GWAS summary data. It corrects genetic correlation estimates derived from cross-trait LD-score regression (LDSC).
}
\details{
The function operates in two modes:

\strong{1. LDSC Mode}: If \code{ldsc_env_path}, \code{ldsc_exe_path}, \code{ld_path}, \code{sumstats1}, and \code{sumstats2} are provided,
the function runs LDSC analysis to compute jackknife blocks estimates for heritability and genetic covariance for the input summary statistics.

\strong{2. Pre-computed Mode}: If paths to jackknife blocks estimates (\code{h21block_path}, \code{h22block_path}, \code{gen_cov_path})
are provided, the function performs downstream analysis without running LDSC.

The downstream analysis computes the heritabilities, genetic covariance, and genetic correlation
with and without corrections, along with their standard errors and p-values using pseudovalues.
}
\examples{
# Example using LDSC Mode
#hvpsum(ldsc_env_path = "path/to/env//ldsc/bin/python", ldsc_exe_path = "path/to/ldsc.py", 
#      ld_path = "path/to/ld", sumstats1 = "path/to/file1.sumstats.gz", 
#       sumstats2 = "path/to/file2.sumstats.gz", tau = 0.5, intern = TRUE)

# Example using Precomputed Mode
#hvpsum(h21block_path = "path/to/file1.hsq1.delete", h22block_path = "path/to/file2.hsq2.delete", 
#       gen_cov_path = "path/to/file3.gencov.delete", tau = 0.5, intern = FALSE)

}
